+++ src/main/java/com/develhope/spring/features/rentals/RentalModel.java
@@ -16,8 +16,8 @@ public class RentalModel {
     private Long id;
     private OffsetDateTime startOfRental;
     private OffsetDateTime endOfRental;
-    private Float dailyCostRental;
-    private Float totalCostRental;
+    private Integer dailyCostRental;
+    private Integer totalCostRental;
     private PaymentStatus status;
     private VehicleEntity vehicleEntity;
     private UserEntity userEntity;
