+++ src/main/java/com/develhope/spring/features/rentals/dto/RentalResponse.java
@@ -14,8 +14,8 @@ public class RentalResponse {
     private Long id;
     private OffsetDateTime startOfRental;
     private OffsetDateTime endOfRental;
-    private Float dailyCostRental;
-    private Float totalCostRental;
+    private Integer dailyCostRental;
+    private Integer totalCostRental;
     private PaymentStatus status;
     private VehicleEntity vehicleEntity;
     private UserEntity userEntity;
